Fix Xcode project VST3 build target
This was a pain - the VST3 target didn't build out of source; it
still thought it was in the iPlug2 Examples folder.

I deleted and manually re-added all VST3_SDK cpp files to the target.
It took a lot of trial and error because there are platform specific
and sample project files mixed in; it's just a mess.